Conflicts:
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.browser.prefs
	.metadata/.plugins/org.eclipse.debug.core/.launches/Tomcat v7.0 Server at localhost.launch
	.metadata/.plugins/org.eclipse.debug.core/.launches/_inicio.jsp.launch
	.metadata/.plugins/org.eclipse.debug.ui/launchConfigurationHistory.xml
	.metadata/.plugins/org.eclipse.jdt.debug.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.jdt.ui/OpenTypeHistory.xml
	.metadata/.plugins/org.eclipse.jdt.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.jst.jsp.core/translators/3706103545.translator
	.metadata/.plugins/org.eclipse.jst.jsp.core/translators/56765979.translator
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/FarmaciaYa/2015/10/41/refactorings.history
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/FarmaciaYa/2015/10/41/refactorings.index
	.metadata/.plugins/org.eclipse.wst.server.core/tmp0/work/Catalina/localhost/FarmaciaYa/org/apache/jsp/cambiocontrasena_jsp.java
	.metadata/.plugins/org.eclipse.wst.server.core/tmp0/work/Catalina/localhost/FarmaciaYa/org/apache/jsp/resultadodebusqueda_jsp.java
	.metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/FarmaciaYa/WEB-INF/classes/META-INF/persistence.xml
	.metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/FarmaciaYa/resultadodebusqueda.jsp
	FarmaciaYa/WebContent/resultadodebusqueda.jsp

+++ FarmaciaYa/src/com/is3/EnvioCorreo.java
@@ -29,7 +29,7 @@ public class EnvioCorreo {
 
 		try {
 			Posicion p = new Posicion();
-			p.getDirCiuByLatLong(new BigDecimal("-34.899124"), new BigDecimal("-56.1454787"));
+			p.getDirCiuByLatLong(new Float("-34.899124"), new Float("-56.1454787"));
 			System.out.println(p.getDireccion());
 			System.out.println(p.getCiudad());
